Move SaveUserDto.ts in SaveUserUseCase.ts and added gitkeep file to keep dto file that will stock later shared dtos
